Refactor Sprites
