+++ lib/dreadnought/core/turret.ex
@@ -14,12 +14,12 @@ defmodule Dreadnought.Core.Turret do
     use Dreadnought.Math
     use Dreadnought.PositionOrientationSize
     use Dreadnought.TypedStruct
-  alias Dreadnought.Core.Sprites
   # TODO replace with use of LinearAlgebra?
   alias Dreadnought.LinearAlgebra.HasCsys
   alias Dreadnought.LinearAlgebra.Vector
   # TODO needed with the use above?
   alias Dreadnought.Math
+  alias Dreadnought.Sprite
 
   # *** *******************************
   # *** TYPES
@@ -28,7 +28,7 @@ defmodule Dreadnought.Core.Turret do
 
   typedstruct enforce: true do
     field :id, integer()
-    field :sprite, Sprites.t
+    field :sprite, Sprite.t
     field :max_ccw_angle, degrees :: number()
     field :max_rotation, positive_degrees :: number()
     field :rest_angle, degrees :: number()
@@ -86,7 +86,7 @@ defmodule Dreadnought.Core.Turret do
   def gun_barrel_vector(%__MODULE__{sprite: sprite}) do
     %{x: x} =
       sprite
-      |> Sprites.mounts
+      |> Sprite.mounts
       |> List.first
       |> position_new
     {x, 0}
