+++ lib/dreadnought/sprite/builder.ex
@@ -1,10 +1,10 @@
 defmodule Dreadnought.Sprite.Builder do
 
     use Dreadnought.Sprite.Spec
-  alias Dreadnought.Core.Sprites
+  alias Dreadnought.Sprite.Importer
 
   def build({function_atom, arg} = sprite_spec) when is_sprite_spec(sprite_spec) do
-    apply(Sprites, function_atom, [arg])
+    apply(Importer, function_atom, [arg])
   end
 
 end
